docker/oanda-candles-to-bq/requirements.txt
requests
pytz
pandas
pandas-gbq
python-dotenv